requirements.txt
black==24.8.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
diff-cover==9.2.0
exceptiongroup==1.2.2
greenlet==3.1.1
importlib-resources==6.4.5
iniconfig==2.1.0
jinja2==3.1.6
MarkupSafe==2.1.5
mypy-extensions==1.0.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlalchemy==2.0.40
sqlfluff==3.3.1
tblib==3.0.0
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.13.1
tzdata==2025.2
zipp==3.20.2
